pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "easyconvio"
version = "1.0.0"
description = "Simple file conversion and transformation for Python"
readme = "README.md"
authors = [{ name = "Joao Feck", email = "joaomfeck@gmail.com" }]
requires-python = ">=3.13"
license = "BSD-3-Clause"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "Pillow",
    "pydub",
    "pypandoc",
    "moviepy",
    "audioop-lts",
]

[project.urls]
Homepage = "https://github.com/jmfeck/easyconvio"
Repository = "https://github.com/jmfeck/easyconvio"
Issues = "https://github.com/jmfeck/easyconvio/issues"

[project.optional-dependencies]
images = ["pillow-heif"]
presentations = ["python-pptx"]
spreadsheets = ["openpyxl", "odfpy", "xlrd"]
vectors = ["cairosvg"]
fonts = ["fonttools", "brotli"]
archives = ["py7zr", "rarfile"]
cad = ["ezdxf", "matplotlib"]
all = [
    "pillow-heif",
    "python-pptx",
    "openpyxl",
    "odfpy",
    "xlrd",
    "cairosvg",
    "fonttools",
    "brotli",
    "py7zr",
    "rarfile",
    "ezdxf",
    "matplotlib",
]
dev = ["pytest", "pytest-cov"]

[tool.setuptools.packages.find]
include = ["easyconvio*"]
